+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * close the session.events chan in StopPoll
 * declare message types here instead of LS
 * Add types for use with the streaming plugin
+* Add types for removing the mountpoints with the streaming plugin